+++ src/schemathesis/__init__.py
@@ -1,2 +1,2 @@
+from .loaders import Parametrizer, from_dict, from_file, from_path, from_uri
 from .models import Case
-from .readers import Parametrizer, from_dict, from_file, from_path, from_uri