+++ test/test_utils.py
@@ -1,0 +1,12 @@
+from schemathesis.utils import is_schemathesis_test
+
+
+def test_is_schemathesis_test(swagger_20):
+    # When a test is wrapped with `parametrize`
+
+    @swagger_20.parametrize()
+    def test():
+        pass
+
+    # Then is should be recognized as a schemathesis test
+    assert is_schemathesis_test(test)